Delete LudoGameFinal.py
Do not want to confuse the TAs that this is the final version of LudoGame.py that I want graded. This was originally going to be the final version of the game for the portfolio, since the class wants an easier to grade version of LudoGame, rather than one that functions exactly how it should as if people are playing the game.